+++ src/main/java/com/kosa/fillinv/global/response/ErrorCode.java
@@ -47,6 +47,7 @@ public enum ErrorCode {
     SCHEDULE_TIME_NOT_FOUND(NOT_FOUND, "SCH07", "해당 스케쥴 시간을 찾을 수 없습니다."),
     SCHEDULE_TIME_MISMATCH(NOT_FOUND, "SCH08", "스케쥴과 스케쥴 시간이 일치하지 않습니다."),
     INVALID_SCHEDULE_STATUS(NOT_FOUND, "SCH09", "스케쥴 상태가 올바르지 않습니다."),
+    INVALID_DATE_FORMAT(NOT_FOUND, "SCH09", "날짜 형식이 올바르지 않습니다."),
 
     // Review Error
     REVIEW_NOT_ALLOWED(BAD_REQUEST, "R01", "리뷰를 작성할 수 없는 상태입니다."),
