+++ src/main/java/com/kosa/fillinv/schedule/controller/ScheduleController.java
@@ -3,10 +3,13 @@ package com.kosa.fillinv.schedule.controller;
 import com.kosa.fillinv.global.response.SuccessResponse;
 import com.kosa.fillinv.global.security.details.CustomMemberDetails;
 import com.kosa.fillinv.schedule.dto.request.ScheduleCreateRequest;
+import com.kosa.fillinv.schedule.dto.request.ScheduleSearchRequest;
 import com.kosa.fillinv.schedule.dto.response.ScheduleDetailResponse;
 import com.kosa.fillinv.schedule.dto.response.ScheduleListResponse;
 import com.kosa.fillinv.schedule.entity.ScheduleStatus;
-import com.kosa.fillinv.schedule.service.ScheduleService;
+import com.kosa.fillinv.schedule.service.ScheduleCreateService;
+import com.kosa.fillinv.schedule.service.ScheduleInquiryService;
+import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springdoc.core.annotations.ParameterObject;
 import org.springframework.data.domain.Page;
@@ -15,10 +18,12 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
@@ -27,7 +32,8 @@ import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 @RequestMapping("/api/v1/schedules")
 public class ScheduleController {
 
-    private final ScheduleService scheduleService;
+    private final ScheduleCreateService scheduleCreateService;
+    private final ScheduleInquiryService scheduleInquiryService;
 
     // 스케쥴 생성
     @PostMapping
@@ -37,7 +43,7 @@ public class ScheduleController {
     ) {
         String memberId = customMemberDetails.memberId();
 
-        String scheduleId = scheduleService.createSchedule(memberId, request);
+        String scheduleId = scheduleCreateService.createSchedule(memberId, request);
 
         // 요청 주소 - ServletUriComponentsBuilder 사용 시 서버 주소가 바뀌더라도 코드를 수정하지 않아도 됨
         // 멘토, 멘티의 스케쥴 상세 보기 주소를 Location 헤더에 담아주기
@@ -52,28 +58,39 @@ public class ScheduleController {
                 .body(SuccessResponse.success(HttpStatus.CREATED));
     }
 
-    // 스케쥴 전체 조회 (GET) - 시간순 정렬 (D-day가 적게 남은 순으로 정렬)
+    // 스케쥴 전체 조회 / 대시보드 - 캘린더 전체 조회 (GET) - 시간순 정렬 (D-day가 적게 남은 순으로 정렬)
+    @GetMapping
+    public ResponseEntity<SuccessResponse<List<ScheduleListResponse>>> getScheduleList(
+            @RequestParam String loginMemberId, // 로그인한 사용자 식별
+            @ModelAttribute ScheduleSearchRequest filter // 필터링 조건 (date, status, title)
+    ) {
+        List<ScheduleListResponse> response = scheduleInquiryService.getScheduleList(loginMemberId, filter);
+
+        return ResponseEntity
+                .ok(SuccessResponse.success(HttpStatus.OK, response));
+    }
 
     // 스케쥴 상세 조회
+    // Ex: GET /api/v1/schedules/1/times/95e3a0e6-e685-4a60-ab63-880031fd4c69
     @GetMapping("/{scheduleId}/times/{scheduleTimeId}")
     public ResponseEntity<SuccessResponse<ScheduleDetailResponse>> getScheduleDetails(
             @PathVariable String scheduleId,
             @PathVariable String scheduleTimeId
     ) {
-        ScheduleDetailResponse response = scheduleService.getScheduleDetail(scheduleId, scheduleTimeId);
+        ScheduleDetailResponse response = scheduleInquiryService.getScheduleDetail(scheduleId, scheduleTimeId);
 
         return ResponseEntity
                 .ok(SuccessResponse.success(HttpStatus.OK, response));
     }
 
     // 상태 일치 스케쥴 조회
-    // Ex: GET /api/v1/schedules/SCH001/status/PAYMENT_PENDING
+    // Ex: GET /api/v1/schedules/1/status/PAYMENT_PENDING
     @GetMapping("/{scheduleId}/status/{status}")
     public ResponseEntity<SuccessResponse<ScheduleListResponse>> getScheduleStatus(
             @PathVariable String scheduleId,
             @PathVariable ScheduleStatus status
     ) {
-        ScheduleListResponse response = scheduleService.getScheduleStatus(scheduleId, status);
+        ScheduleListResponse response = scheduleInquiryService.getScheduleStatus(scheduleId, status);
 
         return ResponseEntity
                 .ok(SuccessResponse.success(HttpStatus.OK, response));
@@ -82,25 +99,25 @@ public class ScheduleController {
     // 스케쥴 상태 변경 (PATCH)
 
     // 멘티 모드: 내 수강 신청 목록 조회 (페이지네이션)
-    // Ex: GET /api/v1/schedules/mentee/MEMBER001?page=0&size=10
+    // Ex: GET /api/v1/schedules/mentee/12?page=0&size=10
     @GetMapping("/mentee/{memberId}") // role=MENTOR or role=MENTEE
     public ResponseEntity<SuccessResponse<Page<ScheduleListResponse>>> getMenteeSchedules(
             @PathVariable String memberId,
             // 기본 10개씩, 생성일자 기준 내림차순(최신순)
             @ParameterObject Pageable pageable
     ) {
-        Page<ScheduleListResponse> responses = scheduleService.getMenteeSchedules(memberId, pageable);
+        Page<ScheduleListResponse> responses = scheduleInquiryService.getMenteeSchedules(memberId, pageable);
         return ResponseEntity.ok(SuccessResponse.success(HttpStatus.OK, responses));
     }
 
     // 멘토 모드: 내 수업 일정 목록 조회 (페이지네이션)
-    // Ex: GET /api/v1/schedules/mentor/MEMBER002?page=0&size=10
+    // Ex: GET /api/v1/schedules/mentor/11?page=0&size=10
     @GetMapping("/mentor/{memberId}") // role=MENTOR or role=MENTEE
     public ResponseEntity<SuccessResponse<Page<ScheduleListResponse>>> getMentorSchedules(
             @PathVariable String memberId,
             @ParameterObject Pageable pageable
     ) {
-        Page<ScheduleListResponse> responses = scheduleService.getMentorSchedules(memberId, pageable);
+        Page<ScheduleListResponse> responses = scheduleInquiryService.getMentorSchedules(memberId, pageable);
         return ResponseEntity.ok(SuccessResponse.success(HttpStatus.OK, responses));
     }
