+++ src/main/java/com/kosa/fillinv/schedule/dto/response/ScheduleListResponse.java
@@ -11,12 +11,12 @@ public record ScheduleListResponse( // 스케쥴 상태 일치 조회
         String menteeNickname,
         Instant startTime,
         ScheduleStatus status,
-        Integer price,
         String lessonType,
-        String optionName
+        String optionName,
+        String userRole // "MENTOR" 또는 "MENTEE"
 ) {
-    public static ScheduleListResponse from(Schedule s, String mentorNickname, String menteeNickname,
-            Instant startTime) {
+    // 역할이 이미 정해진 목록 조회용 (mentee/mentor 전용 API)
+    public static ScheduleListResponse from(Schedule s, String mentorNickname, String menteeNickname, Instant startTime) {
         return new ScheduleListResponse(
                 s.getId(),
                 s.getLessonTitle(),
@@ -24,9 +24,24 @@ public record ScheduleListResponse( // 스케쥴 상태 일치 조회
                 menteeNickname,
                 startTime,
                 s.getStatus(),
-                s.getPrice(),
                 s.getLessonType(),
-                s.getOptionName()
+                s.getOptionName(),
+                null // 역할을 굳이 보내지 않음
+        );
+    }
+
+    // 역할 구분이 필요한 통합 조회용 (캘린더/상세 API)
+    public static ScheduleListResponse from(Schedule s, String mentorNickname, String menteeNickname, Instant startTime, String userRole) {
+        return new ScheduleListResponse(
+                s.getId(),
+                s.getLessonTitle(),
+                mentorNickname,
+                menteeNickname,
+                startTime,
+                s.getStatus(),
+                s.getLessonType(),
+                s.getOptionName(),
+                userRole
         );
     }
 }
